Crate bevy_math

Provides math types and functionality for the Bevy game engine.

The commonly used types are vectors like Vec2 and Vec3, matrices like Mat2, Mat3 and Mat4 and orientation representations like Quat.

Modules

bool

bool vector mask types.

cubic_splines

Provides types for building cubic splines for rendering curves and use with animation easing.

f32

f32 vector, quaternion and matrix types.

f64

f64 vector, quaternion and matrix types.

i32

i32 vector types.

prelude

The bevy_math prelude.

swizzles

Traits adding swizzle methods to all vector types.

u32

u32 vector types.

Structs

Affine2

A 2D affine transform, which can represent translation, rotation, scaling and shear.

Affine3A

A 3D affine transform, which can represent translation, rotation, scaling and shear.

BVec2

A 2-dimensional bool vector mask.

BVec3

A 3-dimensional bool vector mask.

BVec3A

A 3-dimensional SIMD vector mask.

BVec4

A 4-dimensional bool vector mask.

BVec4A

A 4-dimensional SIMD vector mask.

DAffine2

A 2D affine transform, which can represent translation, rotation, scaling and shear.

DAffine3

A 3D affine transform, which can represent translation, rotation, scaling and shear.

DMat2

A 2x2 column major matrix.

DMat3

A 3x3 column major matrix.

DMat4

A 4x4 column major matrix.

DQuat

A quaternion representing an orientation.

DVec2

A 2-dimensional vector.

DVec3

A 3-dimensional vector.

DVec4

A 4-dimensional vector.

IVec2

A 2-dimensional vector.

IVec3

A 3-dimensional vector.

IVec4

A 4-dimensional vector.

Mat2

A 2x2 column major matrix.

Mat3

A 3x3 column major matrix.

Mat3A

A 3x3 column major matrix.

Mat4

A 4x4 column major matrix.

Quat

A quaternion representing an orientation.

Ray

A ray is an infinite line starting at origin, going in direction.

Rect

A rectangle defined by two opposite corners.

UVec2

A 2-dimensional vector.

UVec3

A 3-dimensional vector.

UVec4

A 4-dimensional vector.

Vec2

A 2-dimensional vector.

Vec3

A 3-dimensional vector.

Vec3A

A 3-dimensional vector with SIMD support.

Vec4

A 4-dimensional vector with SIMD support.

Enums

EulerRot

Euler rotation sequences.

Traits

Vec2Swizzles

Vec3Swizzles

Vec4Swizzles

Functions

dmat2

Creates a 2x2 matrix from column vectors.

dmat3

Creates a 3x3 matrix from column vectors.

dmat4

Creates a 4x4 matrix from column vectors.

dquat

Creates a quaternion from x, y, z and w values.

dvec2

Creates a 2-dimensional vector.

dvec3

Creates a 3-dimensional vector.

dvec4

Creates a 4-dimensional vector.

ivec2

Creates a 2-dimensional vector.

ivec3

Creates a 3-dimensional vector.

ivec4

Creates a 4-dimensional vector.

mat2

Creates a 2x2 matrix from column vectors.

mat3

Creates a 3x3 matrix from column vectors.

mat3a

Creates a 3x3 matrix from column vectors.

mat4

Creates a 4x4 matrix from column vectors.

quat

Creates a quaternion from x, y, z and w values.

uvec2

Creates a 2-dimensional vector.

uvec3

Creates a 3-dimensional vector.

uvec4

Creates a 4-dimensional vector.

vec2

Creates a 2-dimensional vector.

vec3

Creates a 3-dimensional vector.

vec3a

Creates a 3-dimensional vector.

vec4

Creates a 4-dimensional vector.