Struct bevy_math::bounding::BoundingCircle

pub struct BoundingCircle {
    pub center: Vec2,
    pub circle: Circle,
}

A bounding circle

Fields
center: Vec2

The center of the bounding circle

circle: Circle

The circle

Implementations

impl BoundingCircle

pub fn new(center: Vec2, radius: f32) -> Self

Constructs a bounding circle from its center and radius.

pub fn from_point_cloud( translation: Vec2, rotation: f32, points: &[Vec2] ) -> BoundingCircle

Computes a BoundingCircle containing the given set of points, transformed by translation and rotation.

The bounding circle is not guaranteed to be the smallest possible.

pub fn radius(&self) -> f32

Get the radius of the bounding circle

pub fn aabb_2d(&self) -> Aabb2d

Computes the smallest Aabb2d containing this BoundingCircle.

pub fn closest_point(&self, point: Vec2) -> Vec2

Finds the point on the bounding circle that is closest to the given point.

If the point is outside the circle, the returned point will be on the perimeter of the circle. Otherwise, it will be inside the circle and returned as is.