bevy_math::bounding
Trait Bounded3d

pub trait Bounded3d {
    // Required methods
    fn aabb_3d(&self, isometry: impl Into<Isometry3d>) -> Aabb3d;
    fn bounding_sphere(&self, isometry: impl Into<Isometry3d>) -> BoundingSphere;
}

A trait with methods that return 3D bounding volumes for a shape.

Required Methods

fn aabb_3d(&self, isometry: impl Into<Isometry3d>) -> Aabb3d

Get an axis-aligned bounding box for the shape translated and rotated by the given isometry.

fn bounding_sphere(&self, isometry: impl Into<Isometry3d>) -> BoundingSphere

Get a bounding sphere for the shape translated and rotated by the given isometry.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.