Trait bevy_math::bounding::IntersectsVolume

pub trait IntersectsVolume<Volume: BoundingVolume> {
    // Required method
    fn intersects(&self, volume: &Volume) -> bool;
}

A trait that generalizes intersection tests against a volume. Intersection tests can be used for a variety of tasks, for example:

Raycasting

Testing for overlap

Checking if an object is within the view frustum of a camera

Required Methods

fn intersects(&self, volume: &Volume) -> bool

Check if a volume intersects with this intersection test