Struct bevy_math::cubic_splines::CubicSegment

pub struct CubicSegment<P: Point> { /* private fields */ }

A segment of a cubic curve, used to hold precomputed coefficients for fast interpolation.

Segments can be chained together to form a longer compound curve.

Implementations

impl<P: Point> CubicSegment<P>

pub fn position(&self, t: f32) -> P

Instantaneous position of a point at parametric value t.

pub fn velocity(&self, t: f32) -> P

Instantaneous velocity of a point at parametric value t.

pub fn acceleration(&self, t: f32) -> P

Instantaneous acceleration of a point at parametric value t.

impl CubicSegment<Vec2>

The CubicSegment<Vec2> can be used as a 2-dimensional easing curve for animation.

The x-axis of the curve is time, and the y-axis is the output value. This struct provides methods for extremely fast solves for y given x.

pub fn new_bezier(p1: impl Into<Vec2>, p2: impl Into<Vec2>) -> Self

Construct a cubic Bezier curve for animation easing, with control points p1 and p2. A cubic Bezier easing curve has control point p0 at (0, 0) and p3 at (1, 1), leaving only p1 and p2 as the remaining degrees of freedom. The first and last control points are fixed to ensure the animation begins at 0, and ends at 1.

This is a very common tool for UI animations that accelerate and decelerate smoothly. For example, the ubiquitous “ease-in-out” is defined as (0.25, 0.1), (0.25, 1.0).

pub fn ease(&self, time: f32) -> f32

Given a time within 0..=1, returns an eased value that follows the cubic curve instead of a straight line. This eased result may be outside the range 0..=1, however it will always start at 0 and end at 1: ease(0) = 0 and ease(1) = 1.

let cubic_bezier = CubicSegment::new_bezier((0.25, 0.1), (0.25, 1.0));
assert_eq!(cubic_bezier.ease(0.0), 0.0);
assert_eq!(cubic_bezier.ease(1.0), 1.0);

How cubic easing works

Easing is generally accomplished with the help of “shaping functions”. These are curves that start at (0,0) and end at (1,1). The x-axis of this plot is the current time of the animation, from 0 to 1. The y-axis is how far along the animation is, also from 0 to 1. You can imagine that if the shaping function is a straight line, there is a 1:1 mapping between the time and how far along your animation is. If the time = 0.5, the animation is halfway through. This is known as linear interpolation, and results in objects animating with a constant velocity, and no smooth acceleration or deceleration at the start or end.

y
│         ●
│       ⬈
│     ⬈    
│   ⬈
│ ⬈
●─────────── x (time)

Using cubic Beziers, we have a curve that starts at (0,0), ends at (1,1), and follows a path determined by the two remaining control points (handles). These handles allow us to define a smooth curve. As time (x-axis) progresses, we now follow the curve, and use the y value to determine how far along the animation is.

y
         ⬈➔●
│      ⬈   
│     ↑      
│     ↑
│    ⬈
●➔⬈───────── x (time)

To accomplish this, we need to be able to find the position y on a curve, given the x value. Cubic curves are implicit parametric functions like B(t) = (x,y). To find y, we first solve for t that corresponds to the given x (time). We use the Newton-Raphson root-finding method to quickly find a value of t that is very near the desired value of x. Once we have this we can easily plug that t into our curve’s position function, to find the y component, which is how far along our animation should be. In other words:

Given time in 0..=1

Use Newton’s method to find a value of t that results in B(t) = (x,y) where x == time

Once a solution is found, use the resulting y value as the final result