Trait bevy_math::cubic_splines::CubicGenerator

pub trait CubicGenerator<P: Point> {
    // Required method
    fn to_curve(&self) -> CubicCurve<P>;
}

Implement this on cubic splines that can generate a curve from their spline parameters.

Required Methods

fn to_curve(&self) -> CubicCurve<P>

Build a CubicCurve by computing the interpolation coefficients for each curve segment.