bevy_math::cubic_splines
Trait RationalGenerator

pub trait RationalGenerator<P: VectorSpace> {
    type Error;

    // Required method
    fn to_curve(&self) -> Result<RationalCurve<P>, Self::Error>;
}

Implement this on cubic splines that can generate a rational cubic curve from their spline parameters.

Required Associated Types

type Error

An error type indicating why construction might fail.

Required Methods

fn to_curve(&self) -> Result<RationalCurve<P>, Self::Error>

Build a RationalCurve by computing the interpolation coefficients for each curve segment.