bevy_math::curve
Module sample_curves

Sample-interpolated curves constructed using the Curve API.

Structs

SampleAutoCurve

A curve that is defined by neighbor interpolation over a set of evenly-spaced samples, interpolated automatically using a particularly well-behaved interpolation.

SampleCurve

A curve that is defined by explicit neighbor interpolation over a set of evenly-spaced samples.

UnevenSampleAutoCurve

A curve that is defined by interpolation over unevenly spaced samples, interpolated automatically using a particularly well-behaved interpolation.

UnevenSampleCurve

A curve that is defined by interpolation over unevenly spaced samples with explicit interpolation.