Module bevy_math::primitives

This module defines primitive shapes. The origin is (0, 0) for 2D primitives and (0, 0, 0) for 3D primitives, unless stated otherwise.

Structs

BoxedPolygon

A polygon with a variable number of vertices, allocated on the heap in a Box<[Vec2]>.

BoxedPolyline2d

A series of connected line segments in 2D space, allocated on the heap in a Box<[Vec2]>.

BoxedPolyline3d

A series of connected line segments in 3D space, allocated on the heap in a Box<[Vec3]>.

Capsule2d

A 2D capsule primitive, also known as a stadium or pill shape.

Capsule3d

A 3D capsule primitive. A three-dimensional capsule is defined as a surface at a distance (radius) from a line

Circle

A circle primitive

Cone

A cone primitive.

ConicalFrustum

A conical frustum primitive. A conical frustum can be created by slicing off a section of a cone.

Cuboid

A cuboid primitive, more commonly known as a box.

Cylinder

A cylinder primitive

Direction2d

A normalized vector pointing in a direction in 2D space

Direction3d

A normalized vector pointing in a direction in 3D space

Ellipse

An ellipse primitive

Line2d

An infinite line along a direction in 2D space.

Line3d

An infinite line along a direction in 3D space.

Plane2d

An unbounded plane in 2D space. It forms a separating surface through the origin, stretching infinitely far

Plane3d

An unbounded plane in 3D space. It forms a separating surface through the origin, stretching infinitely far

Polygon

A polygon with N vertices.

Polyline2d

A series of connected line segments in 2D space.

Polyline3d

A series of connected line segments in 3D space.

Rectangle

A rectangle primitive

RegularPolygon

A polygon where all vertices lie on a circle, equally far apart.

Segment2d

A segment of a line along a direction in 2D space.

Segment3d

A segment of a line along a direction in 3D space.

Sphere

A sphere primitive

Torus

A torus primitive, often representing a ring or donut shape

Triangle2d

A triangle in 2D space

Enums

InvalidDirectionError

An error indicating that a direction is invalid.

TorusKind

The type of torus determined by the minor and major radii

WindingOrder

The winding order for a set of points

Traits

Primitive2d

A marker trait for 2D primitives

Primitive3d

A marker trait for 3D primitives