Struct bevy_math::primitives::RegularPolygon

pub struct RegularPolygon {
    pub circumcircle: Circle,
    pub sides: usize,
}

A polygon where all vertices lie on a circle, equally far apart.

Fields
circumcircle: Circle

The circumcircle on which all vertices lie

sides: usize

The number of sides

Implementations

impl RegularPolygon

pub fn new(circumradius: f32, sides: usize) -> Self

Create a new RegularPolygon from the radius of the circumcircle and a number of sides

Panics

Panics if circumradius is non-positive

pub fn circumradius(&self) -> f32

Get the radius of the circumcircle on which all vertices of the regular polygon lie

pub fn inradius(&self) -> f32

Get the inradius or apothem of the regular polygon. This is the radius of the largest circle that can be drawn within the polygon

pub fn side_length(&self) -> f32

Get the length of one side of the regular polygon

pub fn area(&self) -> f32

Get the area of the regular polygon

pub fn perimeter(&self) -> f32

Get the perimeter of the regular polygon. This is the sum of its sides

pub fn internal_angle_degrees(&self) -> f32

Get the internal angle of the regular polygon in degrees.

This is the angle formed by two adjacent sides with points within the angle being in the interior of the polygon

pub fn internal_angle_radians(&self) -> f32

Get the internal angle of the regular polygon in radians.

This is the angle formed by two adjacent sides with points within the angle being in the interior of the polygon

pub fn external_angle_degrees(&self) -> f32

Get the external angle of the regular polygon in degrees.

This is the angle formed by two adjacent sides with points within the angle being in the exterior of the polygon

pub fn external_angle_radians(&self) -> f32

Get the external angle of the regular polygon in radians.

This is the angle formed by two adjacent sides with points within the angle being in the exterior of the polygon

pub fn vertices(self, rotation: f32) -> impl IntoIterator<Item = Vec2>

Returns an iterator over the vertices of the regular polygon, rotated counterclockwise by the given angle in radians.

With a rotation of 0, a vertex will be placed at the top (0.0, circumradius).