bevy_math
Struct BVec3

#[repr(C, align(1))]
pub struct BVec3 {
    pub x: bool,
    pub y: bool,
    pub z: bool,
}

A 3-dimensional bool vector mask.

Fields
x: booly: boolz: bool
Implementations

impl BVec3

pub const FALSE: BVec3

All false.

pub const TRUE: BVec3

All true.

pub const fn new(x: bool, y: bool, z: bool) -> BVec3

Creates a new vector mask.

pub const fn splat(v: bool) -> BVec3

Creates a vector mask with all elements set to v.

pub const fn from_array(a: [bool; 3]) -> BVec3

Creates a new vector mask from a bool array.

pub fn bitmask(self) -> u32

Returns a bitmask with the lowest 3 bits set from the elements of self.

A true element results in a 1 bit and a false element in a 0 bit. Element x goes into the first lowest bit, element y into the second, etc.

pub fn any(self) -> bool

Returns true if any of the elements are true, false otherwise.

pub fn all(self) -> bool

Returns true if all the elements are true, false otherwise.

pub fn test(&self, index: usize) -> bool

Tests the value at index.

Panics if index is greater than 2.

pub fn set(&mut self, index: usize, value: bool)

Sets the element at index.

Panics if index is greater than 2.