Struct bevy_math::Rot2

pub struct Rot2 {
    pub cos: f32,
    pub sin: f32,
}

A counterclockwise 2D rotation.

Example

use std::f32::consts::PI;

// Create rotations from radians or degrees
let rotation1 = Rot2::radians(PI / 2.0);
let rotation2 = Rot2::degrees(45.0);

// Get the angle back as radians or degrees
assert_eq!(rotation1.as_degrees(), 90.0);
assert_eq!(rotation2.as_radians(), PI / 4.0);

// "Add" rotations together using `*`
assert_relative_eq!(rotation1 * rotation2, Rot2::degrees(135.0));

// Rotate vectors
assert_relative_eq!(rotation1 * Vec2::X, Vec2::Y);

Fields
cos: f32

The cosine of the rotation angle in radians.

This is the real part of the unit complex number representing the rotation.

sin: f32

The sine of the rotation angle in radians.

This is the imaginary part of the unit complex number representing the rotation.

Implementations

impl Rot2

pub const IDENTITY: Self = _

No rotation.

pub const PI: Self = _

A rotation of π radians.

pub const FRAC_PI_2: Self = _

A counterclockwise rotation of π/2 radians.

pub const FRAC_PI_3: Self = _

A counterclockwise rotation of π/3 radians.

pub const FRAC_PI_4: Self = _

A counterclockwise rotation of π/4 radians.

pub const FRAC_PI_6: Self = _

A counterclockwise rotation of π/6 radians.

pub const FRAC_PI_8: Self = _

A counterclockwise rotation of π/8 radians.

pub fn radians(radians: f32) -> Self

Creates a Rot2 from a counterclockwise angle in radians.

pub fn degrees(degrees: f32) -> Self

Creates a Rot2 from a counterclockwise angle in degrees.

pub fn from_sin_cos(sin: f32, cos: f32) -> Self

Creates a Rot2 from the sine and cosine of an angle in radians.

The rotation is only valid if sin * sin + cos * cos == 1.0.

Panics

Panics if sin * sin + cos * cos != 1.0 when the glam_assert feature is enabled.

pub fn as_radians(self) -> f32

Returns the rotation in radians in the (-pi, pi] range.

pub fn as_degrees(self) -> f32

Returns the rotation in degrees in the (-180, 180] range.

pub const fn sin_cos(self) -> (f32, f32)

Returns the sine and cosine of the rotation angle in radians.

pub fn length(self) -> f32

Computes the length or norm of the complex number used to represent the rotation.

The length is typically expected to be 1.0. Unexpectedly denormalized rotations can be a result of incorrect construction or floating point error caused by successive operations.

pub fn length_squared(self) -> f32

Computes the squared length or norm of the complex number used to represent the rotation.

This is generally faster than Rot2::length(), as it avoids a square root operation.

The length is typically expected to be 1.0. Unexpectedly denormalized rotations can be a result of incorrect construction or floating point error caused by successive operations.

pub fn length_recip(self) -> f32

Computes 1.0 / self.length().

For valid results, self must not have a length of zero.

pub fn try_normalize(self) -> Option<Self>

Returns self with a length of 1.0 if possible, and None otherwise.

None will be returned if the sine and cosine of self are both zero (or very close to zero), or if either of them is NaN or infinite.

Note that Rot2 should typically already be normalized by design. Manual normalization is only needed when successive operations result in accumulated floating point error, or if the rotation was constructed with invalid values.

pub fn normalize(self) -> Self

Returns self with a length of 1.0.

Note that Rot2 should typically already be normalized by design. Manual normalization is only needed when successive operations result in accumulated floating point error, or if the rotation was constructed with invalid values.

Panics

Panics if self has a length of zero, NaN, or infinity when debug assertions are enabled.

pub fn is_finite(self) -> bool

Returns true if the rotation is neither infinite nor NaN.

pub fn is_nan(self) -> bool

Returns true if the rotation is NaN.

pub fn is_normalized(self) -> bool

Returns whether self has a length of 1.0 or not.

Uses a precision threshold of approximately 1e-4.

pub fn is_near_identity(self) -> bool

Returns true if the rotation is near Rot2::IDENTITY.

pub fn angle_between(self, other: Self) -> f32

Returns the angle in radians needed to make self and other coincide.

pub fn inverse(self) -> Self

Returns the inverse of the rotation. This is also the conjugate of the unit complex number representing the rotation.

pub fn nlerp(self, end: Self, s: f32) -> Self

Performs a linear interpolation between self and rhs based on the value s, and normalizes the rotation afterwards.

When s == 0.0, the result will be equal to self. When s == 1.0, the result will be equal to rhs.

This is slightly more efficient than slerp, and produces a similar result when the difference between the two rotations is small. At larger differences, the result resembles a kind of ease-in-out effect.

If you would like the angular velocity to remain constant, consider using slerp instead.

Details

nlerp corresponds to computing an angle for a point at position s on a line drawn between the endpoints of the arc formed by self and rhs on a unit circle, and normalizing the result afterwards.

Note that if the angles are opposite like 0 and π, the line will pass through the origin, and the resulting angle will always be either self or rhs depending on s. If s happens to be 0.5 in this case, a valid rotation cannot be computed, and self will be returned as a fallback.

Example

let rot1 = Rot2::IDENTITY;
let rot2 = Rot2::degrees(135.0);

let result1 = rot1.nlerp(rot2, 1.0 / 3.0);
assert_eq!(result1.as_degrees(), 28.675055);

let result2 = rot1.nlerp(rot2, 0.5);
assert_eq!(result2.as_degrees(), 67.5);

pub fn slerp(self, end: Self, s: f32) -> Self

Performs a spherical linear interpolation between self and end based on the value s.

This corresponds to interpolating between the two angles at a constant angular velocity.

When s == 0.0, the result will be equal to self. When s == 1.0, the result will be equal to rhs.

If you would like the rotation to have a kind of ease-in-out effect, consider using the slightly more efficient nlerp instead.

Example

let rot1 = Rot2::IDENTITY;
let rot2 = Rot2::degrees(135.0);

let result1 = rot1.slerp(rot2, 1.0 / 3.0);
assert_eq!(result1.as_degrees(), 45.0);

let result2 = rot1.slerp(rot2, 0.5);
assert_eq!(result2.as_degrees(), 67.5);