Struct bevy_math::URect

#[repr(C)]
pub struct URect {
    pub min: UVec2,
    pub max: UVec2,
}

A rectangle defined by two opposite corners.

The rectangle is axis aligned, and defined by its minimum and maximum coordinates, stored in URect::min and URect::max, respectively. The minimum/maximum invariant must be upheld by the user when directly assigning the fields, otherwise some methods produce invalid results. It is generally recommended to use one of the constructor methods instead, which will ensure this invariant is met, unless you already have the minimum and maximum corners.

Fields
min: UVec2

The minimum corner point of the rect.

max: UVec2

The maximum corner point of the rect.

Implementations

impl URect

pub fn new(x0: u32, y0: u32, x1: u32, y1: u32) -> Self

Create a new rectangle from two corner points.

The two points do not need to be the minimum and/or maximum corners. They only need to be two opposite corners.

Examples

let r = URect::new(0, 4, 10, 6); // w=10 h=2
let r = URect::new(2, 4, 5, 0); // w=3 h=4

pub fn from_corners(p0: UVec2, p1: UVec2) -> Self

Create a new rectangle from two corner points.

The two points do not need to be the minimum and/or maximum corners. They only need to be two opposite corners.

Examples

// Unit rect from [0,0] to [1,1]
let r = URect::from_corners(UVec2::ZERO, UVec2::ONE); // w=1 h=1
// Same; the points do not need to be ordered
let r = URect::from_corners(UVec2::ONE, UVec2::ZERO); // w=1 h=1

pub fn from_center_size(origin: UVec2, size: UVec2) -> Self

Create a new rectangle from its center and size.

Rounding Behaviour

If the size contains odd numbers they will be rounded down to the nearest whole number.

Panics

This method panics if any of the components of the size is negative or if origin - (size / 2) results in any negatives.

Examples

let r = URect::from_center_size(UVec2::ONE, UVec2::splat(2)); // w=2 h=2
assert_eq!(r.min, UVec2::splat(0));
assert_eq!(r.max, UVec2::splat(2));

pub fn from_center_half_size(origin: UVec2, half_size: UVec2) -> Self

Create a new rectangle from its center and half-size.

Panics

This method panics if any of the components of the half-size is negative or if origin - half_size results in any negatives.

Examples

let r = URect::from_center_half_size(UVec2::ONE, UVec2::ONE); // w=2 h=2
assert_eq!(r.min, UVec2::splat(0));
assert_eq!(r.max, UVec2::splat(2));

pub fn is_empty(&self) -> bool

Check if the rectangle is empty.

Examples

let r = URect::from_corners(UVec2::ZERO, UVec2::new(0, 1)); // w=0 h=1
assert!(r.is_empty());

pub const fn width(&self) -> u32

Rectangle width (max.x - min.x).

Examples

let r = URect::new(0, 0, 5, 1); // w=5 h=1
assert_eq!(r.width(), 5);

pub const fn height(&self) -> u32

Rectangle height (max.y - min.y).

Examples

let r = URect::new(0, 0, 5, 1); // w=5 h=1
assert_eq!(r.height(), 1);

pub fn size(&self) -> UVec2

Rectangle size.

Examples

let r = URect::new(0, 0, 5, 1); // w=5 h=1
assert_eq!(r.size(), UVec2::new(5, 1));

pub fn half_size(&self) -> UVec2

Rectangle half-size.

Rounding Behaviour

If the full size contains odd numbers they will be rounded down to the nearest whole number when calculating the half size.

Examples

let r = URect::new(0, 0, 4, 2); // w=4 h=2
assert_eq!(r.half_size(), UVec2::new(2, 1));

pub fn center(&self) -> UVec2

The center point of the rectangle.

Rounding Behaviour

If the (min + max) contains odd numbers they will be rounded down to the nearest whole number when calculating the center.

Examples

let r = URect::new(0, 0, 4, 2); // w=4 h=2
assert_eq!(r.center(), UVec2::new(2, 1));

pub fn contains(&self, point: UVec2) -> bool

Check if a point lies within this rectangle, inclusive of its edges.

Examples

let r = URect::new(0, 0, 5, 1); // w=5 h=1
assert!(r.contains(r.center()));
assert!(r.contains(r.min));
assert!(r.contains(r.max));

pub fn union(&self, other: Self) -> Self

Build a new rectangle formed of the union of this rectangle and another rectangle.

The union is the smallest rectangle enclosing both rectangles.

Examples

let r1 = URect::new(0, 0, 5, 1); // w=5 h=1
let r2 = URect::new(1, 0, 3, 8); // w=2 h=4
let r = r1.union(r2);
assert_eq!(r.min, UVec2::new(0, 0));
assert_eq!(r.max, UVec2::new(5, 8));

pub fn union_point(&self, other: UVec2) -> Self

Build a new rectangle formed of the union of this rectangle and a point.

The union is the smallest rectangle enclosing both the rectangle and the point. If the point is already inside the rectangle, this method returns a copy of the rectangle.

Examples

let r = URect::new(0, 0, 5, 1); // w=5 h=1
let u = r.union_point(UVec2::new(3, 6));
assert_eq!(u.min, UVec2::ZERO);
assert_eq!(u.max, UVec2::new(5, 6));

pub fn intersect(&self, other: Self) -> Self

Build a new rectangle formed of the intersection of this rectangle and another rectangle.

The intersection is the largest rectangle enclosed in both rectangles. If the intersection is empty, this method returns an empty rectangle (URect::is_empty() returns true), but the actual values of URect::min and URect::max are implementation-dependent.

Examples

let r1 = URect::new(0, 0, 2, 2); // w=2 h=2
let r2 = URect::new(1, 1, 3, 3); // w=2 h=2
let r = r1.intersect(r2);
assert_eq!(r.min, UVec2::new(1, 1));
assert_eq!(r.max, UVec2::new(2, 2));

pub fn inset(&self, inset: i32) -> Self

Create a new rectangle with a constant inset.

The inset is the extra border on all sides. A positive inset produces a larger rectangle, while a negative inset is allowed and produces a smaller rectangle. If the inset is negative and its absolute value is larger than the rectangle half-size, the created rectangle is empty.

Examples

let r = URect::new(4, 4, 6, 6); // w=2 h=2
let r2 = r.inset(1); // w=4 h=4
assert_eq!(r2.min, UVec2::splat(3));
assert_eq!(r2.max, UVec2::splat(7));

let r = URect::new(4, 4, 8, 8); // w=4 h=4
let r2 = r.inset(-1); // w=2 h=2
assert_eq!(r2.min, UVec2::splat(5));
assert_eq!(r2.max, UVec2::splat(7));

pub fn as_rect(&self) -> Rect

Returns self as Rect (f32)

pub fn as_irect(&self) -> IRect

Returns self as IRect (i32)