Struct bevy_math::UVec2

#[repr(transparent)]pub struct UVec2(_);

A 2-dimensional vector.

Implementations

impl UVec2

pub const ZERO: UVec2

All zeroes.

pub const ONE: UVec2

All ones.

pub const X: UVec2

[1, 0]: a unit-length vector pointing along the positive X axis.

pub const Y: UVec2

[0, 1]: a unit-length vector pointing along the positive Y axis.

pub fn new(x: u32, y: u32) -> UVec2

Creates a new vector.

pub const fn unit_x() -> UVec2

👎 Deprecated:
Use Vec2::X instead

Creates a vector with values [x: 1.0, y: 0.0].

pub const fn unit_y() -> UVec2

👎 Deprecated:
Use Vec2::Y instead

Creates a vector with values [x: 0.0, y: 1.0].

pub fn extend(self, z: u32) -> UVec3

Creates a 3D vector from self and the given z value.

pub const fn zero() -> UVec2

👎 Deprecated:
use ZERO constant instead

Creates a vector with all elements set to 0.0.

pub const fn one() -> UVec2

👎 Deprecated:
use ONE constant instead

Creates a vector with all elements set to 1.0.

pub fn splat(v: u32) -> UVec2

Creates a vector with all elements set to v.

pub fn select(mask: BVec2, if_true: UVec2, if_false: UVec2) -> UVec2

Creates a vector from the elements in if_true and if_false, selecting which to use for each element of self.

A true element in the mask uses the corresponding element from if_true, and false uses the element from if_false.

pub fn dot(self, other: UVec2) -> u32

Computes the dot product of self and other.

pub fn min(self, other: UVec2) -> UVec2

Returns a vector containing the mininum values for each element of self and other.

In other words this computes [self.x.max(other.x), self.y.max(other.y), ..].

pub fn max(self, other: UVec2) -> UVec2

Returns a vector containing the maximum values for each element of self and other.

In other words this computes [self.x.max(other.x), self.y.max(other.y), ..].

pub fn clamp(self, min: UVec2, max: UVec2) -> UVec2

Component-wise clamping of values, similar to [std::f32::clamp].

Each element in min must be less-or-equal to the corresponing element in max.

If the glam-assert feature is enabled, the function will panic if the contract is not met, otherwise the behavior is undefined.

pub fn min_element(self) -> u32

Returns the horizontal minimum of self.

In other words this computes min(x, y, ..).

pub fn max_element(self) -> u32

Returns the horizontal maximum of self.

In other words this computes max(x, y, ..).

pub fn cmpeq(self, other: UVec2) -> BVec2

Returns a vector mask containing the result of a == comparison for each element of self and other.

In other words, this computes [self.x == other.x, self.y == other.y, ..] for all elements.

pub fn cmpne(self, other: UVec2) -> BVec2

Returns a vector mask containing the result of a != comparison for each element of self and other.

In other words this computes [self.x != other.x, self.y != other.y, ..] for all elements.

pub fn cmpge(self, other: UVec2) -> BVec2

Returns a vector mask containing the result of a >= comparison for each element of self and other.

In other words this computes [self.x >= other.x, self.y >= other.y, ..] for all elements.

pub fn cmpgt(self, other: UVec2) -> BVec2

Returns a vector mask containing the result of a > comparison for each element of self and other.

In other words this computes [self.x > other.x, self.y > other.y, ..] for all elements.

pub fn cmple(self, other: UVec2) -> BVec2

Returns a vector mask containing the result of a <= comparison for each element of self and other.

In other words this computes [self.x <= other.x, self.y <= other.y, ..] for all elements.

pub fn cmplt(self, other: UVec2) -> BVec2

Returns a vector mask containing the result of a < comparison for each element of self and other.

In other words this computes [self.x < other.x, self.y < other.y, ..] for all elements.

pub fn from_slice_unaligned(slice: &[u32]) -> UVec2

Creates a vector from the first N values in slice.

Panics

Panics if slice is less than N elements long.

pub fn write_to_slice_unaligned(self, slice: &mut [u32])

Writes the elements of self to the first N elements in slice.

Panics

Panics if slice is less than N elements long.

pub fn as_f32(&self) -> Vec2

Casts all elements of self to f32.

pub fn as_f64(&self) -> DVec2

Casts all elements of self to f64.

pub fn as_i32(&self) -> IVec2

Casts all elements of self to i32.