bevy_math
Module sampling

This module contains tools related to random sampling.

To use this, the “rand” feature must be enabled.

Re-exports

pub use mesh_sampling::*;

pub use shape_sampling::*;

pub use standard::*;

Modules

mesh_sampling

Functionality related to random sampling from triangle meshes.

shape_sampling

The ShapeSample trait, allowing random sampling from geometric shapes.

standard

This module holds local implementations of the Distribution trait for Standard, which allow certain Bevy math types (those whose values can be randomly generated without additional input other than an Rng) to be produced using rand’s APIs. It also holds FromRng, an ergonomic extension to that functionality which permits the omission of type annotations.